app/converters/image_converter.py
"""
图片转换器
"""
import asyncio
from typing import Dict, Any, Optional
from loguru import logger

try:
    from PIL import Image
    import pytesseract
except ImportError as e:
    logger.warning(f"图片转换器依赖库未安装: {e}")


class ImageConverter:
    """图片转换器类"""
    
    def __init__(self):
        """初始化图片转换器"""
        self.supported_formats = ['pdf', 'docx', 'xlsx', 'pptx', 'jpg', 'png', 'gif', 'bmp', 'tiff']
    
    async def image_to_pdf(
        self, 
        input_path: str, 
        output_path: str, 
        options: Optional[Dict[str, Any]] = None
    ) -> Dict[str, Any]:
        """
        图片转换为 PDF
        
        Args:
            input_path: 输入文件路径
            output_path: 输出文件路径
            options: 转换选项
            
        Returns:
            转换结果
        """
        try:
            # TODO: 实现图片转 PDF 功能
            return {
                'success': False,
                'error': '图片转 PDF 功能暂未实现'
            }
        except Exception as e:
            logger.error(f"图片转 PDF 失败: {e}")
            return {
                'success': False,
                'error': f'图片转 PDF 失败: {str(e)}'
            }
    
    async def image_to_office(
        self, 
        input_path: str, 
        output_path: str, 
        target_format: str, 
        options: Optional[Dict[str, Any]] = None
    ) -> Dict[str, Any]:
        """
        图片转换为 Office 文档
        
        Args:
            input_path: 输入文件路径
            output_path: 输出文件路径
            target_format: 目标格式
            options: 转换选项
            
        Returns:
            转换结果
        """
        try:
            # TODO: 实现图片转 Office 功能
            return {
                'success': False,
                'error': '图片转 Office 功能暂未实现'
            }
        except Exception as e:
            logger.error(f"图片转 Office 失败: {e}")
            return {
                'success': False,
                'error': f'图片转 Office 失败: {str(e)}'
            }
    
    async def image_to_image(
        self, 
        input_path: str, 
        output_path: str, 
        target_format: str, 
        options: Optional[Dict[str, Any]] = None
    ) -> Dict[str, Any]:
        """
        图片格式间转换
        
        Args:
            input_path: 输入文件路径
            output_path: 输出文件路径
            target_format: 目标格式
            options: 转换选项
            
        Returns:
            转换结果
        """
        try:
            # 打开图片
            with Image.open(input_path) as image:
                # 处理图片模式
                if image.mode in ('RGBA', 'LA'):
                    # 转换为 RGB
                    background = Image.new('RGB', image.size, (255, 255, 255))
                    background.paste(image, mask=image.split()[-1] if image.mode == 'RGBA' else None)
                    image = background
                elif image.mode != 'RGB':
                    image = image.convert('RGB')
                
                # 设置图片质量
                quality = options.get('quality', 'medium') if options else 'medium'
                if quality == 'high':
                    save_quality = 95
                elif quality == 'low':
                    save_quality = 60
                else:
                    save_quality = 80
                
                # 调整图片尺寸
                if options and options.get('output_size') and options.get('output_size') != 'original':
                    image = self._resize_image(image, options['output_size'])
                
                # 保存图片
                # 处理格式映射
                save_format = target_format.upper()
                if save_format == 'JPG':
                    save_format = 'JPEG'
                elif save_format == 'TIFF':
                    save_format = 'TIFF'
                
                # 设置保存参数
                save_kwargs = {}
                if save_format == 'JPEG':
                    save_kwargs['quality'] = save_quality
                    save_kwargs['optimize'] = True
                elif save_format == 'PNG':
                    save_kwargs['optimize'] = True
                elif save_format == 'GIF':
                    save_kwargs['optimize'] = True
                
                image.save(output_path, save_format, **save_kwargs)
                
                return {
                    'success': True,
                    'message': f'图片转 {target_format.upper()} 成功'
                }
                
        except Exception as e:
            logger.error(f"图片格式转换失败: {e}")
            return {
                'success': False,
                'error': f'图片格式转换失败: {str(e)}'
            }
    
    async def pdf_to_image(
        self, 
        input_path: str, 
        output_path: str, 
        target_format: str, 
        options: Optional[Dict[str, Any]] = None
    ) -> Dict[str, Any]:
        """
        PDF 转换为图片
        
        Args:
            input_path: 输入文件路径
            output_path: 输出文件路径
            target_format: 目标格式
            options: 转换选项
            
        Returns:
            转换结果
        """
        try:
            # 这里调用 PDF 转换器的方法
            from .pdf_converter import PDFConverter
            pdf_converter = PDFConverter()
            return await pdf_converter.pdf_to_image(input_path, output_path, target_format, options)
        except Exception as e:
            logger.error(f"PDF 转图片失败: {e}")
            return {
                'success': False,
                'error': f'PDF 转图片失败: {str(e)}'
            }
    
    async def office_to_image(
        self, 
        input_path: str, 
        output_path: str, 
        target_format: str, 
        options: Optional[Dict[str, Any]] = None
    ) -> Dict[str, Any]:
        """
        Office 文档转换为图片
        
        Args:
            input_path: 输入文件路径
            output_path: 输出文件路径
            target_format: 目标格式
            options: 转换选项
            
        Returns:
            转换结果
        """
        try:
            # TODO: 实现 Office 转图片功能
            return {
                'success': False,
                'error': 'Office 转图片功能暂未实现'
            }
        except Exception as e:
            logger.error(f"Office 转图片失败: {e}")
            return {
                'success': False,
                'error': f'Office 转图片失败: {str(e)}'
            }
    
    def _resize_image(self, image: Image.Image, output_size: str) -> Image.Image:
        """调整图片尺寸"""
        if output_size == 'A4':
            # A4 尺寸: 210mm x 297mm, 300 DPI
            target_width = int(210 * 300 / 25.4)  # mm to pixels
            target_height = int(297 * 300 / 25.4)
        elif output_size == 'letter':
            # Letter 尺寸: 8.5" x 11", 300 DPI
            target_width = int(8.5 * 300)
            target_height = int(11 * 300)
        else:
            return image
        
        # 保持宽高比
        image.thumbnail((target_width, target_height), Image.Resampling.LANCZOS)
        return image
    
    def get_image_info(self, input_path: str) -> Dict[str, Any]:
        """获取图片信息"""
        try:
            with Image.open(input_path) as image:
                return {
                    'format': image.format,
                    'mode': image.mode,
                    'size': image.size,
                    'width': image.width,
                    'height': image.height,
                    'file_size': image.size[0] * image.size[1] * len(image.getbands())
                }
        except Exception as e:
            logger.error(f"获取图片信息失败: {e}")
            return {} 